+++ boards/RAK3183/examples/LoRaWAN_ISM2400/src/lorawan_app.h
@@ -6,6 +6,8 @@ typedef struct {
     uint8_t dev_eui[8];
     uint8_t join_eui[8];
     uint8_t app_key[16];
+		uint8_t class;
+	  uint8_t dr;
 } LoRaWAN_Params;
 
 extern LoRaWAN_Params lora_params;