+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog  
 
+# [1.0.1] - 2023-7-30  
+## Added   
+AT+CLASS - Set/Get class (0-CLASSA 2-CLASSC)
+AT+DR - Set/Get datarate (0-5)
+AT+TCONF - Set/Get RF test config
+AT+TTX - RF test tx 
+AT+TRX - RF test rx continuously receive mode
+AT+TRXNOP - RF test terminate an ongoing continuous rx mode 
+## Fixed  
+
 # [1.0.0] - 2023-7-9  
 ## Added   
 Added AT instruction frame